BoostMatch/system/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
langdetect==1.0.9
deep-translator==1.11.4
sentence-transformers==2.2.2
scikit-learn==1.3.1
joblib==1.3.2
xgboost==2.0.0
numpy==1.24.3
pandas==2.0.3
textstat==0.7.3
vaderSentiment==3.3.2
spacy==3.6.1
openai-whisper==20230314
yt-dlp==2023.10.13
validators==0.22.0
torch==2.0.1
